Add bundling scripts
The current version considers only the cjs builds. This commit adapts
the traspile (and adds the bundling process) to create esm and umd
based bundles. Furthermore the unbundled versions of cjs and esm are
considered here.

Following scripts were added:

* build:esm
* build:cjs
* build:bundles
* build:declarations
* build:all

Package.json was also adapted and considers the cjs and esm module
systems.